requirements.txt
yt-dlp
openai-whisper
pytesseract
Pillow
scenedetect[opencv]
rapidfuzz
pyvis
